Fix rename modules and imports
