+++ src/autogpt_plugins/text_parser/parser_plugin.py
@@ -1,10 +1,10 @@
 import os
-from text_parser import parser_map
+import parsers
 
 def parse_text(filename: str) -> str:
     """
     Parse text from file and return the results as a string.
     """
     file_extension = os.path.splitext(filename)[1].lower()
-    parser = parser_map.get(file_extension)
+    parser = parsers.parser_map.get(file_extension)
     return parser.parse(filename)
