+++ src/autogpt_plugins/text_parser/parsers.py
