+++ src/autogpt_plugins/text_parser/test_parsers.py
@@ -9,7 +9,7 @@ from bs4 import BeautifulSoup
 from xml.etree import ElementTree
 from pypdf import PdfWriter
 from reportlab.pdfgen import canvas
-from text_parser_plugin import parse_text
+from parser_plugin import parse_text
 
 plain_text_str = "Hello, world!"
